--- a/mage_ai/services/spark/spark.py
+++ b/mage_ai/services/spark/spark.py
@@ -3,6 +3,13 @@ from typing import List
 
 from mage_ai.services.spark.config import SparkConfig
 
+try:
+    from pyspark.conf import SparkConf
+    from pyspark.sql import SparkSession
+    SPARK_ENABLED = True
+except Exception:
+    SPARK_ENABLED = False
+
 
 def get_file_names(jars: List) -> List:
     """
@@ -111,8 +118,8 @@ def get_spark_session(spark_config: SparkConfig):
     Returns:
         SparkSession: The Spark session.
     """
-    from pyspark.conf import SparkConf
-    from pyspark.sql import SparkSession
+    if not SPARK_ENABLED:
+        raise Exception('Spark is not supported in current environment.')
 
     if spark_config:
         active_session = SparkSession.getActiveSession()