--- a/mage_ai/settings/models/configuration_option.py
+++ b/mage_ai/settings/models/configuration_option.py
@@ -15,7 +15,7 @@ from mage_ai.data_preparation.models.pipeline import Pipeline
 from mage_ai.data_preparation.shared.utils import get_template_vars
 from mage_ai.settings.platform import project_platform_activated
 from mage_ai.settings.repo import get_repo_path
-from mage_ai.shared.files import get_full_file_paths_containing_item
+from mage_ai.shared.files import get_full_file_paths_containing_multi_items
 from mage_ai.shared.hash import merge_dict
 from mage_ai.shared.models import BaseDataClass
 
@@ -94,10 +94,16 @@ class ConfigurationOption(BaseDataClass):
                     project_path_only = block.project_path
 
             if OptionType.PROJECTS == option_type or OptionType.PROFILES == option_type:
-                project_full_paths = get_full_file_paths_containing_item(
-                  repo_path,
-                  lambda x: x.startswith('dbt_project.y'),
+                config_full_paths = get_full_file_paths_containing_multi_items(
+                    repo_path,
+                    dict(
+                        project=lambda x: x.startswith('dbt_project.y'),
+                        profile=lambda x: x.startswith('profiles.y'),
+                    ),
+                    exclude_hidden_dir=True,
                 )
+                project_full_paths = config_full_paths['project']
+                profile_full_paths_init = config_full_paths['profile']
 
                 if project_path_only is not None:
                     project_full_paths2 = []
@@ -108,11 +114,6 @@ class ConfigurationOption(BaseDataClass):
                         except ValueError:
                             pass
                     project_full_paths = project_full_paths2
-
-                profile_full_paths_init = get_full_file_paths_containing_item(
-                  repo_path,
-                  lambda x: x.startswith('profiles.y'),
-                )
 
                 profile_full_paths = []
                 profile_full_paths_by_project = {}