# Description
<!-- Please include a summary of the change and which issue is fixed.
Please also include relevant motivation and context.
List any dependencies that are required for this change.
-->
* `/api/pipelines/[pipeline_uuid]/configuration_options?configuration_type=dbt&option_type=projects&resource_type=Block`
  * Combine two directory traverse into one
  * Skip hidden folder traverse
  * Local testing: reduce time from 430ms to 160ms
* `/api/pipelines/[pipeline_uuid]?include_block_pipelines=true&includes_outputs=true`
  * Not repeatedly read block cache.
    * Local testing: reduce time from 580ms to 300ms
  * More efficiently detect whether spark is enabled.
    * Local testing: reduce time from 300ms to 200ms

# How Has This Been Tested?
<!-- Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce.
-->

- [x] Tested locally. See results in description.


# Checklist
- [x] The PR is tagged with proper labels (bug, enhancement, feature, documentation)
- [x] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] If new documentation has been added, relative paths have been added to the appropriate section of `docs/mint.json`

cc:
<!-- Optionally mention someone to let them know about this pull request -->